lint and tests loading
